server/requirements.txt
# FastAPI and web server
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Redis
redis==5.0.1
hiredis==2.3.2

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Voice Pipeline
pipecat-ai==0.0.13
openai-whisper==20231117
TTS==0.22.0

# LLM
openai==1.3.5
anthropic==0.7.7
httpx==0.25.2

# MCP (Model Context Protocol)
mcp==0.1.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
email-validator==2.1.0

# Logging
structlog==23.2.0

# CORS
python-cors==1.0.0

